Refactor Zulip monitor into smaller lifecycle-focused modules
Refactored `src/zulip/monitor.ts` by extracting logic into the following new modules:
- `src/zulip/text-utils.ts`: Text processing and normalization.
- `src/zulip/auth.ts`: Authorization and allowlist logic.
- `src/zulip/media-utils.ts`: Media saving and attachment downloading.
- `src/zulip/reactions.ts`: Standardized reaction handling.
- `src/zulip/bootstrap.ts`: Monitor initialization and account resolution.
- `src/zulip/polling.ts`: Core polling loop and error handling.
- `src/zulip/reply-handler.ts`: Reply dispatching and typing indicators.

This refactor reduces the size and complexity of `monitor.ts` while preserving all existing behaviors, logs, and security practices. Tested with `npm run check`.
