+++ src/zulip/text-utils.ts
@@ -1,0 +1,75 @@
+export const DEFAULT_ONCHAR_PREFIXES = [">", "!"];
+
+/**
+ * Strips HTML tags and unescapes common HTML entities from Zulip message content.
+ */
+export function stripHtmlToText(html: string): string {
+  return html
+    .replace(/<[^>]+>/g, "")
+    .replace(/&lt;/g, "<")
+    .replace(/&gt;/g, ">")
+    .replace(/&amp;/g, "&")
+    .replace(/&quot;/g, '"')
+    .replace(/@\*\*([^*]+)\*\*/g, "@$1")
+    .trim();
+}
+
+/**
+ * Removes a mention of the bot from the text.
+ */
+export function normalizeMention(text: string, mention: string | undefined): string {
+  if (!mention) {
+    return text.trim();
+  }
+  const escaped = mention.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
+  const re = new RegExp(`@${escaped}\\b`, "gi");
+  return text.replace(re, " ").replace(/\s+/g, " ").trim();
+}
+
+/**
+ * Resolves the prefixes used for "onchar" chat mode.
+ */
+export function resolveOncharPrefixes(prefixes: string[] | undefined): string[] {
+  const cleaned = prefixes?.map((entry) => entry.trim()).filter(Boolean) ?? DEFAULT_ONCHAR_PREFIXES;
+  return cleaned.length > 0 ? cleaned : DEFAULT_ONCHAR_PREFIXES;
+}
+
+/**
+ * Checks if the text starts with any of the given prefixes and strips it if so.
+ */
+export function stripOncharPrefix(
+  text: string,
+  prefixes: string[],
+): { triggered: boolean; stripped: string } {
+  const trimmed = text.trimStart();
+  for (const prefix of prefixes) {
+    if (!prefix) {
+      continue;
+    }
+    if (trimmed.startsWith(prefix)) {
+      return {
+        triggered: true,
+        stripped: trimmed.slice(prefix.length).trimStart(),
+      };
+    }
+  }
+  return { triggered: false, stripped: text };
+}
+
+/**
+ * Extracts a topic directive (e.g. [[zulip_topic: Topic Name]]) from the text.
+ */
+export function extractZulipTopicDirective(text: string): { text: string; topic?: string } {
+  const match = text.match(/^\s*\[\[zulip_topic:\s*([^\]]+?)\s*\]\]\s*/i);
+  if (!match) {
+    return { text };
+  }
+  const topic = match[1]?.trim();
+  if (!topic) {
+    return { text: text.slice(match[0].length).trimStart() };
+  }
+  return {
+    text: text.slice(match[0].length).trimStart(),
+    topic,
+  };
+}
